+++ NEWS.md
@@ -13,6 +13,7 @@ all additive predictor terms specified in the same formula. (#1492)
 and incompatible with the newly implemented global shrinkage prior framework.
 * No longer support multiple deprecated prior options for categorical and
 multivariate models after around 3 years of deprecation. (#1420)
+* Deprecate argument `newdata` of `get_refmodel.brmsfit()`. (#1502)
 
 ### Bug Fixes
 
@@ -95,6 +96,7 @@ over-dispersed and zero-inflated binomial response models thanks to Hayden
 Rabel. (#1319 & #1311)
 * Display `ppd_*` plots in `pp_check` via argument `prefix`. (#1313)
 * Support the `log` link in binomial and beta type families. (#1316)
+* Support **projpred**'s augmented-data projection. (#1292, #1294)
 
 ### Other changes